Move game model classes/interfaces from store into new src/model directory
Remove unused games vuex module
