+++ frontend/src/model/game/card.ts
@@ -29,13 +29,7 @@ export class Card {
 export type Discard = {
     [color: string]: Card[];
 };
+
 export type CardsInPlay = {
     [color: string]: Card[];
 };
-
-export interface GameData {
-    id: number;
-    discard: Discard;
-    player1: any;
-    player2: any;
-}
