+++ frontend/src/store/index.ts
@@ -9,7 +9,6 @@ import './modules/alert';
 import { AuthState } from './modules/auth';
 import { AccountState } from './modules/account';
 import { AlertState } from './modules/alert';
-import { GameStoreState } from '@/store/modules/game';
 
 Vue.use(Vuex);
 
@@ -17,7 +16,6 @@ export interface RootState {
     account: AccountState;
     auth: AuthState;
     alert: AlertState;
-    game: GameStoreState;
 }
 
 const buildStore = () => getStoreBuilder<RootState>().vuexStore();
