--- frontend/src/store/modules/game/index.ts
@@ -1,7 +1,0 @@
-import * as model from './model';
-import * as moves from './moves';
-
-export default {
-    model,
-    moves,
-};
