--- frontend/src/store/modules/games.ts
@@ -1,1 +1,0 @@
-export default {};
