+++ frontend/src/utils.ts
@@ -1,4 +1,4 @@
-import { Color } from '@/store/modules/game/model';
+import { Color } from '@/model/game/card';
 
 export function getColorEnumValues(): Color[] {
     return Object.keys(Color).map(key => Color[key]);
